- feedback has been addressed - rename bredcrumb folder - set last link as active - rearange styles for better readability - delete useless example file
