+++ src/components/breadcrumbs/__docs__/Breadcrumbs.mdx
@@ -22,7 +22,6 @@ const breadCrumbList = [
   {
     label: 'Catalog',
     link: '/',
-    active: true,
   },
   {
     label: 'View Complaint',
