+++ src/components/index.ts
@@ -17,3 +17,4 @@ export * from "./checkboxInput";
 export * from "./tableHeader";
 export * from "./tooltip";
 export * from "./dropzone";
+export * from "./breadcrumbs";
